requirements.txt
Flask==2.3.2
flask_cors==4.0.1
supabase==1.0.4
Flask-JWT-Extended==4.5.2
flasgger==0.9.7.1
flask-socketio==5.3.6
Flask-APScheduler==1.13.1
requests==2.32.2
aiohttp==3.9.5
asgiref==3.8.1
pillow==10.4.0
imagehash==4.3.1